+++ exporters/trace/jaeger/jaeger_test.go
@@ -148,7 +148,7 @@ func TestNewExportPipeline(t *testing.T) {
 			if tc.testSpanSampling {
 				_, span := tp.Tracer("jaeger test").Start(context.Background(), tc.name)
 				spanCtx := span.SpanContext()
-				assert.Equal(t, tc.spanShouldBeSampled, spanCtx.TraceFlags().IsSampled())
+				assert.Equal(t, tc.spanShouldBeSampled, spanCtx.IsSampled())
 				span.End()
 			}
 		})